+++ src/main/java/tmtd/event/attendance/dto/AttendanceMarkRequest.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.Positive;
 /** Yêu cầu đánh dấu điểm danh */
 public record AttendanceMarkRequest(
     @NotNull @Positive Long eventId,
-    @NotNull @Positive Integer studentId,
+    @NotNull @Positive Long studentId,
     /** "QR" hoặc "MANUAL" (tùy chọn, mặc định QR) */
     String method
 ) {}
