+++ src/main/java/tmtd/event/events/dto/EventCreateRequest.java
@@ -15,7 +15,7 @@ public record EventCreateRequest(
                 LocalDate startDate,
                 LocalDate endDate,
                 String venue,
-                Integer totalSeats,
+                Long totalSeats,
                 String mainImageUrl // <-- thêm
 ) {
 }
