+++ src/main/java/tmtd/event/registrations/dto/RegistrationCreateRequest.java
@@ -10,12 +10,12 @@ public class RegistrationCreateRequest {
 
     // server tự gắn từ token -> KHÔNG validate NotNull ở đây, và ẩn khỏi client
     @JsonIgnore
-    private Integer studentId;
+    private Long studentId;
 
     // getters & setters
     public Long getEventId() { return eventId; }
     public void setEventId(Long eventId) { this.eventId = eventId; }
 
-    public Integer getStudentId() { return studentId; }
-    public void setStudentId(Integer studentId) { this.studentId = studentId; }
+    public Long getStudentId() { return studentId; }
+    public void setStudentId(Long studentId) { this.studentId = studentId; }
 }
